feat: seperate line motion
