+++ crates/motiongfx_vello/src/vello_motion/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod circle_motion;
+pub(crate) mod line_motion;
 pub(crate) mod rect_motion;
